database, update AutoMapper for messages to return, new messages controller.
Create new Dto for message to return.
New method in MessagesController for GetMessagesForUser - via query params we are choose from frontend what kind of box we would like to see: Outbox / Inbox.
This params with userId ( our id ) go to GetMessagesForUser method in appRepo, here messages are taken from database due to param ( Out or In messages ), then messages are returning with pagination ( messages are also ordered ).
Then the messages are mapped to Dto to return and then finally returned.